Robot
Another thing in python that prompts the user for an integer. This
integer sets the parameters for a 2n by 2n square. The robot then moves
in random directions until it reaches n. If by n it does reach, it
should output the number of steps it took. There is also an optional
function that prints where the robot is after each step.